Fix bug (reducers reinitializing values)
Previously, reducers will return the default (initialized) values when
the action type doesn't trigger any of its cases, which is incorrect.
This commit fixes this so that reducers will only do so when received
state is undefined (initialization), but will return the received state
in every other case (no more reinitialization of values).